Added request number report for each json object
Only in json mode request numbers are shown.

+++ interface.h
@@ -61,7 +61,7 @@ struct in_ev {
   int refcnt;
   int error;
   int fd;
-  long long int uid;
+  int uid;
 };
